+++ gluecodium/src/test/resources/smoke/defaults_const/output/android/com/example/smoke/EnumCollectionDefaults.java
@@ -3,6 +3,7 @@
  */
 package com.example.smoke;
 import android.support.annotation.NonNull;
+import com.example.HashMapBuilder;
 import com.example.fire.Enum1;
 import com.example.fire.Enum2;
 import com.example.fire.Enum3;
@@ -15,8 +16,6 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
 public final class EnumCollectionDefaults {
     @NonNull
     public List<Enum1> listField;
@@ -27,6 +26,6 @@ public final class EnumCollectionDefaults {
     public EnumCollectionDefaults() {
         this.listField = new ArrayList<>(Arrays.asList(Enum1.DISABLED));
         this.setField = new EnumSet<>(Arrays.asList(Enum2.DISABLED));
-        this.mapField = new HashMap<>(Stream.of(new AbstractMap.SimpleEntry<>(Enum3.DISABLED, Enum4.DISABLED)).collect(Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue)));
+        this.mapField = new HashMapBuilder<Enum3, Enum4>().put(Enum3.DISABLED, Enum4.DISABLED).build();
     }
 }