+++ gluecodium/src/test/resources/smoke/defaults_const/output/android/com/example/smoke/EnumCollectionDefaultsExternal.java
@@ -3,6 +3,7 @@
  */
 package com.example.smoke;
 import android.support.annotation.NonNull;
+import com.example.HashMapBuilder;
 import java.util.AbstractMap;
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -11,8 +12,6 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
 public final class EnumCollectionDefaultsExternal {
     @NonNull
     public List<foo.AlienEnum1> listField;
@@ -23,6 +22,6 @@ public final class EnumCollectionDefaultsExternal {
     public EnumCollectionDefaultsExternal() {
         this.listField = new ArrayList<>(Arrays.asList(foo.AlienEnum1.DISABLED));
         this.setField = new EnumSet<>(Arrays.asList(foo.AlienEnum2.DISABLED));
-        this.mapField = new HashMap<>(Stream.of(new AbstractMap.SimpleEntry<>(foo.AlienEnum3.DISABLED, foo.AlienEnum4.DISABLED)).collect(Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue)));
+        this.mapField = new HashMapBuilder<foo.AlienEnum3, foo.AlienEnum4>().put(foo.AlienEnum3.DISABLED, foo.AlienEnum4.DISABLED).build();
     }
 }