+++ gluecodium/src/test/resources/smoke/dates/output/dart/lib/src/smoke/dates.dart
@@ -5,6 +5,7 @@ import 'package:library/src/_token_cache.dart' as __lib;
 import 'package:library/src/builtin_types__conversion.dart';
 import 'package:library/src/generic_types__conversion.dart';
 abstract class Dates {
+
   DateTime dateMethod(DateTime input);
   DateTime? nullableDateMethod(DateTime? input);
   DateTime get dateProperty;
@@ -15,9 +16,7 @@ abstract class Dates {
 class Dates_DateStruct {
   DateTime dateField;
   DateTime? nullableDateField;
-  Dates_DateStruct._(this.dateField, this.nullableDateField);
-  Dates_DateStruct(DateTime dateField)
-    : dateField = dateField, nullableDateField = null;
+  Dates_DateStruct(this.dateField, this.nullableDateField);
 }
 // Dates_DateStruct "private" section, not exported.
 final _smokeDatesDatestructCreateHandle = __lib.catchArgumentError(() => __lib.nativeLibrary.lookupFunction<
@@ -48,7 +47,7 @@ Dates_DateStruct smokeDatesDatestructFromFfi(Pointer<Void> handle) {
   final _dateFieldHandle = _smokeDatesDatestructGetFielddateField(handle);
   final _nullableDateFieldHandle = _smokeDatesDatestructGetFieldnullableDateField(handle);
   try {
-    return Dates_DateStruct._(
+    return Dates_DateStruct(
       dateFromFfi(_dateFieldHandle),
       dateFromFfiNullable(_nullableDateFieldHandle)
     );
@@ -103,6 +102,7 @@ final _smokeDatesReleaseHandle = __lib.catchArgumentError(() => __lib.nativeLibr
   >('library_smoke_Dates_release_handle'));
 class Dates$Impl extends __lib.NativeBase implements Dates {
   Dates$Impl(Pointer<Void> handle) : super(handle);
+
   @override
   DateTime dateMethod(DateTime input) {
     final _dateMethodFfi = __lib.catchArgumentError(() => __lib.nativeLibrary.lookupFunction<Uint64 Function(Pointer<Void>, Int32, Uint64), int Function(Pointer<Void>, int, int)>('library_smoke_Dates_dateMethod__Date'));